requirements.txt
streamlit==1.30.0
langchain==0.1.17
langchain-community==0.0.36
psycopg2-binary==2.9.7
pgvector==0.2.4
sentence-transformers==2.6.1
aiohttp==3.9.1
python-dotenv==1.0.0
tqdm==4.66.1
google-generativeai==0.3.1
pypdf==3.17.1
azure-storage-blob==12.15.0
azure-identity==1.15.0
gunicorn==21.2.0
transformers==4.40.2
huggingface_hub==0.23.0
sqlalchemy==2.0.23
